Add range tree supporting search in x-axis, y-axis and arbitrary half-plane
Thoroughly tested. I spent many combined litres of blood, sweat and tears on this
In the end, I'm actually surprised that it worked.
It seems to work, anyway.
Actually, what if it doesn't work, but always appear to work? Does that mean it doesn't work?
If humans are incapable of distinguishing its good-doing with its bad-doing, then why would we even
care to rule out all possibilities for bugs?
I certainly hope Professor Karger doesn't see this